Move `infer_handle_*` to separate header
