+++ libcaf_core/caf/blocking_actor.hpp
@@ -46,6 +46,8 @@ class blocking_actor
     : public extend<local_actor, blocking_actor>::
              with<mixin::sync_sender<blocking_response_handle_tag>::impl> {
 public:
+  using behavior_type = behavior;
+
   blocking_actor();
 
   ~blocking_actor();
